Complete Clean Architecture

+++ internal/kitchen/infrastructure/ai/optimizer_impl.go
@@ -32,51 +32,51 @@ func (s *OptimizerServiceImpl) OptimizeWorkflow(ctx context.Context, orders []*d
 
 	// For now, optimize the first order as an example
 	order := orders[0]
-	
+
 	s.logger.WithField("order_id", order.ID()).Info("Optimizing workflow")
 
 	optimization := domain.NewWorkflowOptimization(order.ID())
-	
+
 	// Analyze required stations
 	requiredStations := order.GetRequiredStations()
-	
+
 	// Create optimized steps based on required stations
 	stepID := 1
 	totalTime := int32(0)
-	
+
 	for _, stationType := range requiredStations {
 		step := &domain.WorkflowStep{
-			StepID:        fmt.Sprintf("step_%d", stepID),
-			StationType:   stationType,
-			EstimatedTime: s.estimateStationTime(stationType, order),
-			RequiredSkill: s.getRequiredSkillForStation(stationType),
-			Dependencies:  s.getDependencies(stepID, requiredStations),
+			StepID:         fmt.Sprintf("step_%d", stepID),
+			StationType:    stationType,
+			EstimatedTime:  s.estimateStationTime(stationType, order),
+			RequiredSkill:  s.getRequiredSkillForStation(stationType),
+			Dependencies:   s.getDependencies(stepID, requiredStations),
 			CanParallelize: s.canParallelize(stationType),
 		}
-		
+
 		optimization.AddStep(step)
 		totalTime += step.EstimatedTime
 		stepID++
 	}
-	
+
 	optimization.CalculateEstimatedTime()
-	
+
 	// Calculate efficiency gain (simplified)
 	baselineTime := int32(len(requiredStations) * 120) // 2 minutes per station baseline
 	if baselineTime > 0 {
 		optimization.EfficiencyGain = float32(baselineTime-totalTime) / float32(baselineTime) * 100
 	}
-	
+
 	// Add recommendations
 	optimization.AddRecommendation("Prioritize parallel processing where possible")
 	optimization.AddRecommendation("Ensure staff specialization matches station requirements")
-	
+
 	s.logger.WithFields(map[string]interface{}{
 		"order_id":        order.ID(),
 		"estimated_time":  optimization.EstimatedTime,
 		"efficiency_gain": optimization.EfficiencyGain,
 	}).Info("Workflow optimization completed")
-	
+
 	return optimization, nil
 }
 
@@ -85,18 +85,18 @@ func (s *OptimizerServiceImpl) PredictPreparationTime(ctx context.Context, order
 	s.logger.WithField("order_id", order.ID()).Info("Predicting preparation time")
 
 	baseTime := int32(60) // 1 minute base time
-	
+
 	// Add time based on items
 	for _, item := range order.Items() {
 		itemTime := s.estimateItemTime(item)
 		baseTime += itemTime * item.Quantity()
 	}
-	
+
 	// Add complexity factor based on number of required stations
 	requiredStations := order.GetRequiredStations()
 	complexityFactor := float32(len(requiredStations)) * 0.2
 	adjustedTime := float32(baseTime) * (1.0 + complexityFactor)
-	
+
 	// Add priority adjustment
 	switch order.Priority() {
 	case domain.OrderPriorityUrgent:
@@ -106,16 +106,16 @@ func (s *OptimizerServiceImpl) PredictPreparationTime(ctx context.Context, order
 	case domain.OrderPriorityLow:
 		adjustedTime *= 1.2 // 20% slower for low priority
 	}
-	
+
 	predictedTime := int32(adjustedTime)
-	
+
 	s.logger.WithFields(map[string]interface{}{
 		"order_id":       order.ID(),
 		"predicted_time": predictedTime,
 		"base_time":      baseTime,
 		"complexity":     complexityFactor,
 	}).Info("Preparation time predicted")
-	
+
 	return predictedTime, nil
 }
 
@@ -131,14 +131,14 @@ func (s *OptimizerServiceImpl) AllocateStaff(ctx context.Context, orders []*doma
 		LoadBalance:     make(map[string]float32),
 		Recommendations: []string{},
 	}
-	
+
 	// Sort orders by priority
 	sortedOrders := make([]*domain.KitchenOrder, len(orders))
 	copy(sortedOrders, orders)
 	sort.Slice(sortedOrders, func(i, j int) bool {
 		return sortedOrders[i].Priority() > sortedOrders[j].Priority()
 	})
-	
+
 	// Sort staff by availability and skill
 	availableStaff := make([]*domain.Staff, 0)
 	for _, s := range staff {
@@ -146,24 +146,24 @@ func (s *OptimizerServiceImpl) AllocateStaff(ctx context.Context, orders []*doma
 			availableStaff = append(availableStaff, s)
 		}
 	}
-	
+
 	sort.Slice(availableStaff, func(i, j int) bool {
 		return availableStaff[i].GetWorkload() < availableStaff[j].GetWorkload()
 	})
-	
+
 	// Allocate staff to orders
 	for _, order := range sortedOrders {
 		if order.Status() != domain.OrderStatusPending {
 			continue
 		}
-		
+
 		requiredStations := order.GetRequiredStations()
 		bestStaff := s.findBestStaffForOrder(order, availableStaff, requiredStations)
-		
+
 		if bestStaff != nil {
 			estimatedTime, _ := s.PredictPreparationTime(ctx, order)
 			efficiency := bestStaff.GetEfficiencyForStation(requiredStations[0])
-			
+
 			orderAllocation := domain.NewStaffOrderAllocation(
 				bestStaff.ID(),
 				order.ID(),
@@ -172,12 +172,12 @@ func (s *OptimizerServiceImpl) AllocateStaff(ctx context.Context, orders []*doma
 				efficiency,
 				"Optimal skill match",
 			)
-			
+
 			allocation.Allocations = append(allocation.Allocations, orderAllocation)
 			allocation.LoadBalance[bestStaff.ID()] = bestStaff.GetWorkload()
 		}
 	}
-	
+
 	// Calculate utilization rate
 	if len(staff) > 0 {
 		totalWorkload := float32(0)
@@ -186,7 +186,7 @@ func (s *OptimizerServiceImpl) AllocateStaff(ctx context.Context, orders []*doma
 		}
 		allocation.UtilizationRate = totalWorkload / float32(len(staff))
 	}
-	
+
 	// Add recommendations
 	if allocation.UtilizationRate > 0.8 {
 		allocation.Recommendations = append(allocation.Recommendations, "Consider adding more staff - high utilization detected")
@@ -194,19 +194,19 @@ func (s *OptimizerServiceImpl) AllocateStaff(ctx context.Context, orders []*doma
 	if len(allocation.Allocations) < len(orders) {
 		allocation.Recommendations = append(allocation.Recommendations, "Some orders could not be allocated - check staff availability")
 	}
-	
+
 	s.logger.WithFields(map[string]interface{}{
 		"allocations_count": len(allocation.Allocations),
 		"utilization_rate":  allocation.UtilizationRate,
 	}).Info("Staff allocation completed")
-	
+
 	return allocation, nil
 }
 
 // OptimizeStaffSchedule optimizes staff schedule
 func (s *OptimizerServiceImpl) OptimizeStaffSchedule(ctx context.Context, staff []*domain.Staff, timeWindow time.Duration) (*application.StaffScheduleOptimization, error) {
 	s.logger.WithFields(map[string]interface{}{
-		"staff_count":  len(staff),
+		"staff_count": len(staff),
 		"time_window": timeWindow.String(),
 	}).Info("Optimizing staff schedule")
 
@@ -215,16 +215,16 @@ func (s *OptimizerServiceImpl) OptimizeStaffSchedule(ctx context.Context, staff 
 		Recommendations: []string{},
 		CreatedAt:       time.Now(),
 	}
-	
+
 	// Create shifts for each staff member
 	for _, s := range staff {
 		if !s.IsAvailable() {
 			continue
 		}
-		
+
 		// Assign to their primary specialization
 		primaryStation := s.Specializations()[0]
-		
+
 		shift := &application.StaffShift{
 			StaffID:   s.ID(),
 			StartTime: time.Now(),
@@ -232,17 +232,17 @@ func (s *OptimizerServiceImpl) OptimizeStaffSchedule(ctx context.Context, staff 
 			Station:   primaryStation,
 			Load:      s.GetWorkload(),
 		}
-		
+
 		optimization.Schedule[s.ID()] = shift
 	}
-	
+
 	// Calculate efficiency gain (simplified)
 	optimization.EfficiencyGain = 15.0 // 15% efficiency gain estimate
 	optimization.CostReduction = 10.0  // 10% cost reduction estimate
-	
+
 	optimization.Recommendations = append(optimization.Recommendations, "Balance workload across all staff members")
 	optimization.Recommendations = append(optimization.Recommendations, "Consider cross-training for better flexibility")
-	
+
 	return optimization, nil
 }
 
@@ -258,13 +258,13 @@ func (s *OptimizerServiceImpl) OptimizeEquipmentUsage(ctx context.Context, equip
 		Recommendations: []string{},
 		CreatedAt:       time.Now(),
 	}
-	
+
 	// Allocate equipment to orders
 	for _, eq := range equipment {
 		if !eq.IsAvailable() {
 			continue
 		}
-		
+
 		// Find orders that need this equipment type
 		var suitableOrders []string
 		for _, order := range orders {
@@ -276,7 +276,7 @@ func (s *OptimizerServiceImpl) OptimizeEquipmentUsage(ctx context.Context, equip
 				}
 			}
 		}
-		
+
 		if len(suitableOrders) > 0 {
 			allocation := &application.EquipmentAllocation{
 				EquipmentID: eq.ID(),
@@ -285,17 +285,17 @@ func (s *OptimizerServiceImpl) OptimizeEquipmentUsage(ctx context.Context, equip
 				EndTime:     time.Now().Add(2 * time.Hour), // 2 hour window
 				Load:        eq.GetUtilizationRate(),
 			}
-			
+
 			optimization.Allocations[eq.ID()] = allocation
 		}
 	}
-	
+
 	// Calculate utilization gain
 	optimization.UtilizationGain = 20.0 // 20% utilization improvement estimate
-	
+
 	optimization.Recommendations = append(optimization.Recommendations, "Schedule maintenance during low-demand periods")
 	optimization.Recommendations = append(optimization.Recommendations, "Monitor equipment efficiency scores regularly")
-	
+
 	return optimization, nil
 }
 
@@ -303,14 +303,14 @@ func (s *OptimizerServiceImpl) OptimizeEquipmentUsage(ctx context.Context, equip
 func (s *OptimizerServiceImpl) PredictEquipmentLoad(ctx context.Context, equipment *domain.Equipment, timeWindow time.Duration) (float32, error) {
 	// Simplified prediction based on current load and time window
 	currentLoad := equipment.GetUtilizationRate()
-	
+
 	// Predict slight increase during peak hours
 	hours := timeWindow.Hours()
 	if hours > 0 && hours <= 4 {
 		// Peak hours - increase load
-		return math.Min(float64(currentLoad*1.3), 1.0), nil
+		return float32(math.Min(float64(currentLoad*1.3), 1.0)), nil
 	}
-	
+
 	// Off-peak hours - maintain or slightly decrease
 	return currentLoad * 0.9, nil
 }
@@ -327,18 +327,18 @@ func (s *OptimizerServiceImpl) PredictCapacity(ctx context.Context, timeWindow t
 		Recommendations:  []string{},
 		PredictedAt:      time.Now(),
 	}
-	
+
 	// Calculate predicted orders based on time window
 	hours := timeWindow.Hours()
 	prediction.PredictedOrders = int32(hours * 10) // 10 orders per hour estimate
-	
+
 	prediction.CapacityGap = prediction.RequiredCapacity - prediction.CurrentCapacity
-	
+
 	if prediction.CapacityGap > 0 {
 		prediction.Recommendations = append(prediction.Recommendations, "Consider adding temporary staff")
 		prediction.Recommendations = append(prediction.Recommendations, "Optimize current workflows")
 	}
-	
+
 	return prediction, nil
 }
 
@@ -356,27 +356,27 @@ func (s *OptimizerServiceImpl) AnalyzeBottlenecks(ctx context.Context) (*applica
 		},
 		AnalyzedAt: time.Now(),
 	}
-	
+
 	// Example bottlenecks
 	bottlenecks := []*application.Bottleneck{
 		{
-			Type:       "equipment",
-			ResourceID: "espresso_machine_1",
-			Severity:   0.7,
-			Impact:     "High utilization causing delays",
+			Type:        "equipment",
+			ResourceID:  "espresso_machine_1",
+			Severity:    0.7,
+			Impact:      "High utilization causing delays",
 			Suggestions: []string{"Add backup equipment", "Schedule maintenance"},
 		},
 		{
-			Type:       "staff",
-			ResourceID: "barista_1",
-			Severity:   0.5,
-			Impact:     "Overloaded during peak hours",
+			Type:        "staff",
+			ResourceID:  "barista_1",
+			Severity:    0.5,
+			Impact:      "Overloaded during peak hours",
 			Suggestions: []string{"Add staff during peak hours", "Cross-train other staff"},
 		},
 	}
-	
+
 	analysis.Bottlenecks = bottlenecks
-	
+
 	return analysis, nil
 }
 
@@ -393,20 +393,20 @@ func (s *OptimizerServiceImpl) AnalyzePerformance(ctx context.Context, period *a
 		Recommendations: []*application.Recommendation{},
 		AnalyzedAt:      time.Now(),
 	}
-	
+
 	// Example metrics
 	analysis.Metrics["efficiency"] = 0.85
 	analysis.Metrics["throughput"] = 45.0
 	analysis.Metrics["quality"] = 0.92
-	
+
 	// Example trends
 	analysis.Trends["efficiency"] = []float32{0.80, 0.82, 0.85, 0.87, 0.85}
 	analysis.Trends["throughput"] = []float32{40.0, 42.0, 45.0, 47.0, 45.0}
-	
+
 	// Example insights
 	analysis.Insights = append(analysis.Insights, "Efficiency has improved by 5% over the period")
 	analysis.Insights = append(analysis.Insights, "Throughput shows steady growth with occasional dips")
-	
+
 	return analysis, nil
 }
 
@@ -438,7 +438,7 @@ func (s *OptimizerServiceImpl) GenerateRecommendations(ctx context.Context, cont
 			CreatedAt:   time.Now(),
 		},
 	}
-	
+
 	return recommendations, nil
 }
 
@@ -446,25 +446,25 @@ func (s *OptimizerServiceImpl) GenerateRecommendations(ctx context.Context, cont
 
 func (s *OptimizerServiceImpl) estimateStationTime(stationType domain.StationType, order *domain.KitchenOrder) int32 {
 	baseTime := map[domain.StationType]int32{
-		domain.StationTypeEspresso: 90,  // 1.5 minutes
-		domain.StationTypeGrinder:  30,  // 30 seconds
-		domain.StationTypeSteamer:  60,  // 1 minute
-		domain.StationTypeAssembly: 45,  // 45 seconds
+		domain.StationTypeEspresso: 90, // 1.5 minutes
+		domain.StationTypeGrinder:  30, // 30 seconds
+		domain.StationTypeSteamer:  60, // 1 minute
+		domain.StationTypeAssembly: 45, // 45 seconds
 	}
-	
+
 	if time, exists := baseTime[stationType]; exists {
 		// Adjust for order complexity
 		complexity := float32(order.GetTotalQuantity()) * 0.1
 		return int32(float32(time) * (1.0 + complexity))
 	}
-	
+
 	return 60 // Default 1 minute
 }
 
 func (s *OptimizerServiceImpl) estimateItemTime(item *domain.OrderItem) int32 {
 	// Base time per item type (simplified)
 	baseTime := int32(30) // 30 seconds per item
-	
+
 	// Adjust based on complexity (number of requirements)
 	complexity := len(item.Requirements())
 	return baseTime + int32(complexity*10)
@@ -477,11 +477,11 @@ func (s *OptimizerServiceImpl) getRequiredSkillForStation(stationType domain.Sta
 		domain.StationTypeSteamer:  7.0, // High skill
 		domain.StationTypeAssembly: 6.0, // Medium-high skill
 	}
-	
+
 	if skill, exists := skillRequirements[stationType]; exists {
 		return skill
 	}
-	
+
 	return 5.0 // Default medium skill
 }
 
@@ -490,7 +490,7 @@ func (s *OptimizerServiceImpl) getDependencies(stepID int, stations []domain.Sta
 	if stepID == 1 {
 		return []string{} // First step has no dependencies
 	}
-	
+
 	return []string{fmt.Sprintf("step_%d", stepID-1)} // Depends on previous step
 }
 
@@ -502,23 +502,23 @@ func (s *OptimizerServiceImpl) canParallelize(stationType domain.StationType) bo
 		domain.StationTypeEspresso: false, // Usually sequential
 		domain.StationTypeSteamer:  false, // Usually sequential
 	}
-	
+
 	if canParallel, exists := parallelizable[stationType]; exists {
 		return canParallel
 	}
-	
+
 	return false
 }
 
 func (s *OptimizerServiceImpl) findBestStaffForOrder(order *domain.KitchenOrder, staff []*domain.Staff, requiredStations []domain.StationType) *domain.Staff {
 	var bestStaff *domain.Staff
 	bestScore := float32(-1)
-	
+
 	for _, s := range staff {
 		if !s.CanAcceptOrder() {
 			continue
 		}
-		
+
 		// Calculate score based on specialization match and workload
 		score := float32(0)
 		for _, station := range requiredStations {
@@ -527,15 +527,15 @@ func (s *OptimizerServiceImpl) findBestStaffForOrder(order *domain.KitchenOrder,
 				break
 			}
 		}
-		
+
 		// Prefer less loaded staff
 		score = score * (1.0 - s.GetWorkload())
-		
+
 		if score > bestScore {
 			bestScore = score
 			bestStaff = s
 		}
 	}
-	
+
 	return bestStaff
 }
